feat: implement select ticket page
